Add new emoji file
I created this emoji file a long time ago when I had the idea for the plugin.

I think it is the most comprehensive set of emoji I could find when I did some research. Then again, it was created in October, so maybe new emoji were added since then.

You may need to perform some further processing to make it useful, such as converting the unicode points to the actual emoji strings. Or maybe this file isn't useful at all. You can decide what do to with it.

Either way, I wanted to upload it somewhere in case it becomes useful in the future.